Attempt at graph with react
The problem with allowing React to be in charge of rendering is that on
each tick the entire React lifecycle is triggered. As a result such an
implementation will not scale.

This commit is retained just to show that it can be done, however a
future commit will use React for structuring only of the visualisation;
we'll leave D3 to take care of the calculations and rendering.